+++ example/constants.jl
@@ -1,0 +1,1 @@
+BIT_WIDTH_CNT = 48;